implement combined scraper

+++ scrapers/project/scrapers/qoo10.py
@@ -1,7 +1,7 @@
 import requests
 from bs4 import BeautifulSoup
-from ecom_scrapers.headers import get_headers
-from ecom_scrapers.scraper import Scraper
+from scrapers.headers import get_headers
+from scrapers.scraper import Scraper
 
 class Qoo10(Scraper):
     url = 'https://www.qoo10.sg'
@@ -35,8 +35,8 @@ class Qoo10(Scraper):
             parsed['pages'] = int(params['pages'])
         if params.get('minPrice') or params.get('maxPrice'):
             minPrice, maxPrice = params.get('minPrice'), params.get('maxPrice')
-            minPrice = float(minPrice) if minPrice else ''
-            maxPrice = float(maxPrice) if maxPrice else ''
+            minPrice = int(minPrice) if minPrice else ''
+            maxPrice = int(maxPrice) if maxPrice else ''
             parsed['f'] += f'pr:{minPrice}-{maxPrice}|'
         if params.get('sort'):
             parsed['s'] = self.sort_choices[params['sort']]
